isit-svc/pyproject.toml
[tool.poetry]
name = "isit-svc"
version = "0.1.0"
description = ""
authors = ["Данилин Даниил <d-danilin@groupprice.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
pydantic = "^2.5.3"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
app = "isit_svc.main:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
